apps/api/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
sqlalchemy>=2.0.0
alembic>=1.13.0
redis>=5.0.0
cohere>=5.0.0
sentence-transformers>=3.0.0
pypdf>=5.0.0
ragas>=0.1.0
pytest>=8.0.0
httpx>=0.27.0
